add some explicit casts

+++ src/rw.c
@@ -118,8 +118,8 @@ static ScsData * read_scs_data(FILE *fin) {
 
 void SCS(write_data)(const ScsData *d, const ScsCone *k) {
   FILE* fout = fopen(d->stgs->write_data_filename, "wb");
-  unsigned long scs_int_sz = sizeof(scs_int);
-  unsigned long scs_float_sz = sizeof(scs_float);
+  unsigned long scs_int_sz = (unsigned long)sizeof(scs_int);
+  unsigned long scs_float_sz = (unsigned long)sizeof(scs_float);
   scs_printf("writing data to %s\n", d->stgs->write_data_filename);
   fwrite(&(scs_int_sz), sizeof(unsigned long), 1, fout);
   fwrite(&(scs_float_sz), sizeof(unsigned long), 1, fout);
@@ -136,17 +136,17 @@ scs_int SCS(read_data)(const char * filename, ScsData ** d, ScsCone ** k) {
   fread(&(file_int_sz), sizeof(unsigned long), 1, fin);
   fread(&(file_float_sz), sizeof(unsigned long), 1, fin);
 
-  if (file_int_sz != sizeof(scs_int)) {
+  if (file_int_sz != (unsigned long)sizeof(scs_int)) {
     scs_printf("Error, sizeof(file int) is %lu, but scs expects sizeof(int) "
         "%lu, scs should be recompiled with correct flags.\n",
-        file_int_sz, sizeof(scs_int));
+        file_int_sz, (unsigned long)sizeof(scs_int));
     fclose(fin);
     RETURN -1;
   }
-  if (file_float_sz != sizeof(scs_float )) {
+  if (file_float_sz != (unsigned long)sizeof(scs_float )) {
     scs_printf("Error, sizeof(file float) is %lu, but scs expects "
         "sizeof(float) %lu, scs should be recompiled with the correct flags.\n",
-        file_float_sz, sizeof(scs_float));
+        file_float_sz, (unsigned long)sizeof(scs_float));
     fclose(fin);
     RETURN -1;
   }
